# Remove hashing from prediction cache