Perf comparison on alexnet:
__With hashing:__
+ 10 clients: 80 qps, mean lat ~= 122 ms
+ 15 clients: 83 qps, mean lat ~= 0.168 ms

__Without hashing__
+ 10 clients: 105 qps, mean lat ~= 95 ms
+ 15 clients: 115 qps, mean lat ~=125 ms ms